pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "openai>=1.82.0",
    "pyyaml>=6.0.2",
    "streamlit>=1.45.1",
]
